+++ changelog.md
@@ -14,6 +14,8 @@ Codion Change Log
 - CalendarPanel, not including time no longer disables the time element of the component dateTime value, only excludes the input fields.
 - DefaultComboBoxSelectionDialogBuilder bug fixed, now uses singular selection dialog title.
 - CalendarPanel.Builder.includeTimeZone() added.
+- DefaultFrameBuilder bug fixed, now always makes sure the frame fits the screen.
+- DefaultComponentDialogBuilder, DefaultListSelectionDialogBuilder and DefaultComboBoxSelectionDialogBuilder bug fixed, now always make sure the dialog fits the screen.
 ### is.codion.plugin.flatlaf.themes
 - Module is.codion.plugin.flatlaf.lookandfeels renamed is.codion.plugin.flatlaf.themes.
 ### is.codion.swing.framework.ui